feat(search): add SearchTable component for improved record search
